220606
Video emulation and 040/060
Misc Falcon related
Gfxcard detection and palette handling
Disk-image cache
WIP (disabled) display mode change on startup
Misc smaller bugs and fixes

+++ BasiliskII/src/macos_util.cpp
@@ -163,7 +163,7 @@ uint32 TimeToMacTime(time_t t)
 	int intervening_leap_days = (a4 - b4) - (a100 - b100) + (a400 - b400);
 	uint32 days = local->tm_yday + 365 * (local->tm_year - 4) + intervening_leap_days;
 	int32 dayofs = -PrefsFindInt32("dayofs");
-	if(dayofs > 0 && dayofs > days)
+	if(dayofs > 0 && ((uint32)dayofs > days))
 		dayofs = days;
 	return local->tm_sec + 60 * (local->tm_min + 60 * (local->tm_hour + 24 * (days - dayofs)));
 }
